# Conflicts:
#	RadeonRays/src/strategy/bvh2lstrategy.cpp
#	RadeonRays/src/strategy/bvhstrategy.cpp
#	RadeonRays/src/strategy/fatbvhstrategy.cpp

+++ App/config_manager.cpp
@@ -57,7 +57,7 @@ void ConfigManager::CreateConfigs(Mode mode, bool interop, std::vector<Config>& 
 	for (int i = 0; i < platforms.size(); ++i)
 	{
 		std::vector<CLWDevice> devices;
-		int startidx = configs.size();
+		int startidx = (int)configs.size();
 
 		for (int d = 0; d < (int)platforms[i].GetDeviceCount(); ++d)
 		{
